Removed all traces of IO from core
Also updated dependencies:
build.sbt: use Scala 2/13/16
core/build.sbt: updated scalatest, flog, scala-parallel..., logback, slf4j (added new)
cats/build.sbt: updated scalatest, tsec, cats-effect, slf4j (new)
