+++ build.sbt
@@ -4,7 +4,7 @@ name := "TableParser"
 
 ThisBuild / version := "1.1.4-SNAPSHOT"
 
-ThisBuild / scalaVersion := "2.13.14"
+ThisBuild / scalaVersion := "2.13.16"
 
 lazy val core = project
