+++ cats/src/test/scala/com/phasmidsoftware/tableparser/cats/table/AnalysisSpec.scala
@@ -1,10 +1,10 @@
 package com.phasmidsoftware.tableparser.cats.table
 
 import cats.effect.IO
+import com.phasmidsoftware.tableparser.cats.util.EvaluateIO.matchIO
 import com.phasmidsoftware.tableparser.core.parse.{RawTableParser, TableParser}
 import com.phasmidsoftware.tableparser.core.table.Column.make
 import com.phasmidsoftware.tableparser.core.table._
-import com.phasmidsoftware.tableparser.core.util.EvaluateIO.matchIO
 import com.phasmidsoftware.tableparser.core.util.FP.{resource, sequence}
 import org.scalatest.concurrent.PatienceConfiguration.Timeout
 import org.scalatest.flatspec.AnyFlatSpec
