+++ core/src/main/scala/com/phasmidsoftware/tableparser/core/write/Writable.scala
@@ -4,7 +4,6 @@
 
 package com.phasmidsoftware.tableparser.core.write
 
-import cats.effect.IO
 import java.io.{File, FileWriter}
 
 /**
@@ -139,14 +138,5 @@ object Writable {
     def writeRaw(o: FileWriter)(x: CharSequence): FileWriter = o.append(x).asInstanceOf[FileWriter]
 
     override def close(o: FileWriter): Unit = o.close()
-  }
-
-  // TESTME (why isn't this used? it should be)
-  def fileWritableIO(file: File): Writable[IO[FileWriter]] = new Writable[IO[FileWriter]] {
-    def unit: IO[FileWriter] = IO(new FileWriter(file))
-
-    def writeRaw(fi: IO[FileWriter])(x: CharSequence): IO[FileWriter] = for (f <- fi) yield f.append(x).asInstanceOf[FileWriter]
-
-    override def close(fi: IO[FileWriter]): Unit = for (f <- fi) yield f.close()
   }
 }
